fix: multimodal exampple
